Fix JAR generation
The previous commit caused non-functional Java Archive files to be
created as the main() method was placed in a class extending JavaFX's
Application.

JAR generation needs to be restored to allow for easier usage of the
project.

Let's rectify JAR generation by moving the main() method into a
different class.
